#401 Add check that points fall within surface
See Issue: #401
Adds check that input points fall within surface
